+++ OneBlink.SDK/models/FormSubmissionEvent.cs
@@ -226,7 +226,7 @@ namespace OneBlink.SDK.Model
             List<string> ccEmail = default(List<string>),
             List<string> bccEmail = default(List<string>),
             bool? includeExternalIdInPdf = null,
-            EndpointConfiguration emailAttachmentsWebhook = null
+            EndpointConfiguration emailAttachmentsEndpoint = null
         )
         {
             FormSubmissionEventConfiguration fseconfig = new FormSubmissionEventConfiguration();
@@ -238,7 +238,7 @@ namespace OneBlink.SDK.Model
             fseconfig.excludedElementIds = excludedElementIds;
             fseconfig.usePagesAsBreaks = usePagesAsBreaks;
             fseconfig.emailTemplate = emailTemplate;
-            fseconfig.emailAttachmentsWebhook = emailAttachmentsWebhook;
+            fseconfig.emailAttachmentsEndpoint = emailAttachmentsEndpoint;
             if (toEmail != default(List<string>))
             {
                 fseconfig.toEmail = toEmail;
@@ -283,7 +283,7 @@ namespace OneBlink.SDK.Model
             List<string> bccEmail = default(List<string>),
             List<string> excludedCSSClasses = default(List<string>),
             bool? includeExternalIdInPdf = null,
-            EndpointConfiguration emailAttachmentsWebhook = null
+            EndpointConfiguration emailAttachmentsEndpoint = null
         )
         {
             FormSubmissionEventConfiguration fseconfig = new FormSubmissionEventConfiguration();
@@ -314,7 +314,7 @@ namespace OneBlink.SDK.Model
                 fseconfig.bccEmail = bccEmail;
             }
             fseconfig.includeExternalIdInPdf = includeExternalIdInPdf;
-            fseconfig.emailAttachmentsWebhook = emailAttachmentsWebhook;
+            fseconfig.emailAttachmentsEndpoint = emailAttachmentsEndpoint;
             FormSubmissionEvent pdfEvent = new FormSubmissionEvent();
             pdfEvent.type = "PDF";
             pdfEvent.configuration = fseconfig;
@@ -341,7 +341,7 @@ namespace OneBlink.SDK.Model
             List<string> toEmail = default(List<string>),
             List<string> ccEmail = default(List<string>),
             List<string> bccEmail = default(List<string>),
-            EndpointConfiguration emailAttachmentsWebhook = null
+            EndpointConfiguration emailAttachmentsEndpoint = null
         )
         {
             FormSubmissionEventConfiguration fseconfig = new FormSubmissionEventConfiguration();
@@ -360,7 +360,7 @@ namespace OneBlink.SDK.Model
             {
                 fseconfig.bccEmail = bccEmail;
             }
-            fseconfig.emailAttachmentsWebhook = emailAttachmentsWebhook;
+            fseconfig.emailAttachmentsEndpoint = emailAttachmentsEndpoint;
             FormSubmissionEvent emailEvent = new FormSubmissionEvent();
             emailEvent.type = "EMAIL";
             emailEvent.configuration = fseconfig;
@@ -385,7 +385,7 @@ namespace OneBlink.SDK.Model
             List<string> toEmail = default(List<string>),
             List<string> ccEmail = default(List<string>),
             List<string> bccEmail = default(List<string>),
-            EndpointConfiguration emailAttachmentsWebhook = null
+            EndpointConfiguration emailAttachmentsEndpoint = null
         )
         {
             FormSubmissionEventConfiguration fseconfig = new FormSubmissionEventConfiguration();
@@ -403,7 +403,7 @@ namespace OneBlink.SDK.Model
             {
                 fseconfig.bccEmail = bccEmail;
             }
-            fseconfig.emailAttachmentsWebhook = emailAttachmentsWebhook;
+            fseconfig.emailAttachmentsEndpoint = emailAttachmentsEndpoint;
             FormSubmissionEvent emailEvent = new FormSubmissionEvent();
             emailEvent.type = "EMAIL";
             emailEvent.configuration = fseconfig;